Move more stuff out of `WindowFunctionEvaluator` class namespace into the `window_function_evaluator` namespace
