+++ src/lib/CMakeLists.txt
@@ -283,9 +283,9 @@ set(
     operators/validate.hpp
     operators/window_function_evaluator.cpp
     operators/window_function_evaluator.hpp
+    operators/window_function_evaluator/util.cpp
+    operators/window_function_evaluator/util.hpp
     operators/window_function_evaluator/window_function_evaluator_traits.hpp
-    operators/window_function_evaluator/relevant_row_information.cpp
-    operators/window_function_evaluator/relevant_row_information.hpp
     optimizer/join_ordering/abstract_join_ordering_algorithm.cpp
     optimizer/join_ordering/abstract_join_ordering_algorithm.hpp
     optimizer/join_ordering/dp_ccp.cpp
